+++ modules/valaxy-addon-live2d/types/index.ts
@@ -1,0 +1,8 @@
+export interface Live2dOptions {
+  global?: boolean
+
+  waifuPath?: string
+  apiPath?: string
+  cdnPath?: string
+  live2dPath?: string
+}